requirements.txt
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.2.21
langchain-community==0.3.0
langchain-core==0.3.9
psycopg2-binary==2.9.10
pandas==2.1.3
pymongo==4.8.0
sentence-transformers==3.1.1
chromadb==0.4.18
plotly==5.17.0
python-dotenv==1.0.1
groq==0.11.0
pydantic==2.9.2
httpx=0.27.2
